WASM host refactor general audio to C64 specific for now.
